+++ resolwe_bio/tools/make_igv_session_archive.py
@@ -1,38 +1,57 @@
 #!/usr/bin/env python3
-# pylint: disable=missing-docstring,invalid-name,import-error
 """Create IGV session for archive."""
 
 import argparse
-from lxml import etree
 import os
 
-parser = argparse.ArgumentParser(description='Create igv session')
+from lxml import etree
 
-parser.add_argument("-f", "--files", nargs='+', required=True, help="List of files")
-parser.add_argument("-g", "--genome", type=str, required=True, help="Genome")
+parser = argparse.ArgumentParser(description="Create igv session.")
+
+parser.add_argument('-f', '--input_file', required=True, help="File with paths to files for IGV")
 
 args = parser.parse_args()
 
-Global = etree.Element(
-    'Global',
-    genome=args.genome,
-    version='8'
-)
 
-Resources = etree.SubElement(Global, 'Resources')
+def make_xml_tree(input_file):
+    """Make xml tree for IGV session."""
+    global_ = etree.Element(  # pylint: disable=no-member
+        'Global',
+        version='3',
+    )
 
-for element in args.files:
-    element_name = os.path.basename(element)
-    element_dir = os.path.dirname(os.path.dirname(element))
-    element_path = os.path.join('..', element)
+    resources = etree.SubElement(global_, 'Resources')  # pylint: disable=no-member
 
-    if element_dir == 'None':
-        element_subdir = os.path.basename(os.path.dirname(element))
-        element_path = os.path.join('..', 'other_data', element_subdir, element_name)
+    with open(input_file, 'r') as tfile:
+        for line in tfile:
+            line = line.rstrip()
+            etree.SubElement(  # pylint: disable=no-member
+                resources,
+                'Resource',
+                name=os.path.basename(line),
+                path=os.path.join('..', line),
+            )
 
-    Resource = etree.SubElement(Resources, 'Resource', name=element_name, path=element_path)
+    doc = etree.tostring(  # pylint: disable=no-member
+        global_,
+        pretty_print=True,
+        xml_declaration=True,
+        encoding='UTF-8',
+    )
+    return doc
 
-doc = etree.tostring(Global, pretty_print=True, xml_declaration=True, encoding='UTF-8')
 
-with open('igv_session.xml', 'wb') as f:
-    f.write(doc)
+def write_xml_file(input_file, doc):
+    """Compose a name and write output file."""
+    with open(input_file.replace('temp_igv.txt', 'igv.xml'), 'wb') as f:
+        f.write(doc)
+
+
+def main():
+    """Invoke when run directly as a program."""
+    doc = make_xml_tree(args.input_file)
+    write_xml_file(args.input_file, doc)
+
+
+if __name__ == "__main__":
+    main()